hierarchical_planner/requirements.txt
google-generativeai
python-dotenv
PyYAML>=6.0 # Specify version for safety, ensure it's installed
pytest
openai
anthropic
